requirements.txt
python-dotenv
python-telegram-bot
google-generativeai
playwright
pillow
feedparser
schedule
requests
